topology: hda-dsp: Remove srcintc module from HDA DSP topology
The main goal of this topology is to handle every device with hda-dsp
configuration on Skylake SST driver.

The module_id value for srcintc module varies by FW branch, which makes
it impossible to have one universal topology with srcintc module for all of them.

Remove srcintc from this topology to make it truly universal.

Signed-off-by: Mateusz Gorski <mateusz.gorski@linux.intel.com>
Signed-off-by: Takashi Iwai <tiwai@suse.de>